feat(ai): hard-delete AI consent instead of revoking it
